Add server setup script & rm old server on install
Previously the new server was installed on top of the earlier one
without first clearing the old one. Now the old one is first removed (if
it is there to begin with) and the server is cloned afterwards.

The server setup script is crucial as the setup service requires it to
be present. On startup the setup script clones the repository and
expects to be able to call the server setup script for completing the
server installation.