Add the base for the serverside
The basic structure of the serverside is the division into two modules.
Module server_handler is responsible for handling the requests whilst
calling the actual functions from server_func. This division allows to
more easily change the implementation of the request handler. For
example, Flask could be omitted and the same interface could be
fulfilled using WSGI. This should be fairly straight forward change as
long as the server_func remains its own component.

Stubs for potentially relevant functions and requests are also added as
reminders of the functionalities.