First deployable version
Change server to be publically reachable. Using host="0.0.0.0" for the
Flask application makes it available to public Internet. The port
number is also changed for avoiding collisions. The script that sets up
the server is updated to start the actual server instead of the test
version. The cookie length is extended in order to make bruteforcing
less viable. The cookie generation needs to make sure no user gets the
same cookie.